camera/pyproject.toml
[tool.poetry]
name = "camera"
version = "0.1.0"
description = ""
authors = ["KennetTech <64315533+KennetTech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
opencv-python = "^4.10.0.84"
python-dotenv = "^1.0.1"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
